chore: wip for setting up routes correctly
