+++ services/passport.js
@@ -2,6 +2,7 @@ const passport = require('passport');
 const JwtStrategy = require('passport-jwt').Strategy;
 const LocalStrategy = require('passport-local').Strategy;
 const User = require('../models/user');
+require('dotenv').config();
 
 const cookieExtractor = (req) => {
   let token = null;
